refactor minor AST
