+++ prql-compiler/src/ast/pl/frame.rs
@@ -1,13 +1,11 @@
-use std::{
-    collections::HashSet,
-    fmt::{Debug, Display, Formatter},
-};
+use std::collections::HashSet;
+use std::fmt::{Debug, Display, Formatter};
 
 use enum_as_inner::EnumAsInner;
 use itertools::{Itertools, Position};
 use serde::{Deserialize, Serialize};
 
-use super::{Expr, Ident};
+use super::Ident;
 
 /// Represents the object that is manipulated by the pipeline transforms.
 /// Similar to a view in a database or a data frame.
@@ -46,24 +44,6 @@ pub enum FrameColumn {
         name: Option<Ident>,
         expr_id: usize,
     },
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ColumnSort<T = Expr> {
-    pub direction: SortDirection,
-    pub column: T,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub enum SortDirection {
-    Asc,
-    Desc,
-}
-
-impl Default for SortDirection {
-    fn default() -> Self {
-        SortDirection::Asc
-    }
 }
 
 impl Display for Frame {
